Stop Localizing Logs
After realizing that a localized log wouldn't be very useful to us, the authors, who are likely to be the only ones using the logs for anything. Also, not running everything through the localizer should increase logging performance.

On second thought , let’s not go to Camelot. ‘Tis a silly place.